move linspace and logspace to tools, fix some warnings
