+++ src/preprocess.rs
@@ -6,6 +6,7 @@ use std::f64::consts::PI;
 use std::path::Path;
 
 use crate::elements::NIElement;
+use crate::tools;
 use crate::Cplx;
 
 pub fn get_num_threads() -> usize {
@@ -149,32 +150,6 @@ fn process_collocation_pts(mesh: &mut mesh_data::Mesh, body_id: usize) ->
     return (eqn_map, cpt_map)
 }
 
-pub fn linspace(start: f64, end: f64, npts: usize) -> Vec<f64> {
-    // evenly-spaced npts between start and end
-    let dx = (end - start) / ((npts - 1) as f64);
-    let mut x = vec![start; npts];
-    for i in 1..(npts-1) {
-        x[i] = start + dx * (i as f64);
-    }
-    x[npts-1] = end;
-    return x;
-}
-
-pub fn logspace(start: f64, end: f64, npts: usize) -> Vec<f64> {
-    // evenly-spaced npts between start and end, evaluated 
-    // such that there are equal intervals between log(start)
-    // and log(end)
-    let start_log = start.log10();
-    let end_log = end.log10();
-    let x_log = linspace(start_log, end_log, npts);
-    let mut x = vec![start; npts];
-    x[npts-1] = end;
-    for i in 1..(npts-1) {
-        x[i] = 10.0_f64.powf(x_log[i]);
-    }
-    return x;
-}
-
 fn process_frequency_list(freq_input: &input_data::FrequencyInput) 
     -> Vec<f64> {
     match freq_input {
@@ -182,10 +157,10 @@ fn process_frequency_list(freq_input: &input_data::FrequencyInput)
             return values.to_vec();
         },
         input_data::FrequencyInput::LinearSpaced { start, end, number } => {
-            return linspace(*start, *end, *number);
+            return tools::linspace(*start, *end, *number);
         },
         input_data::FrequencyInput::LogSpaced { start, end, number } => {
-            return logspace(*start, *end, *number);
+            return tools::logspace(*start, *end, *number);
         }
     }
 }
